Refactoring and disabled link logic for main branch
